compositor|refactor: Decompose the plan consistency assertion
Each plan consistency check is now its own module, so a test exercises one rule directly rather than building a whole plan, running every check, and digging the result out of a thrown error. The assertion composes the checks and throws, and nothing else.

A failing plan raises a failure that can be caught as the shared consistency base, while still reporting its own name and wording.
