+++ packages/compositor/src/plan/__tests__/assertPlanIsConsistent.unit.test.ts
@@ -1,10 +1,14 @@
 import { describe, expect, it } from 'vitest';
 
+import { ConsistencyError } from '../../consistency/ConsistencyError.ts';
 import type { Plan } from '../../schemas/plan-schema.ts';
 import { buildPlan } from '../../test-utils/buildPlan.ts';
 import { requireEntry } from '../../test-utils/requireEntry.ts';
-import { assertPlanIsConsistent, PlanConsistencyError, type PlanViolation } from '../assertPlanIsConsistent.ts';
+import { assertPlanIsConsistent, PlanConsistencyError } from '../assertPlanIsConsistent.ts';
 
+// Each check is covered against its own module under `checks/__tests__/`. What is left here is what only the
+// composition can show: that a check is wired in at all, that the checks do not interfere, and that one run reports
+// every violation it found.
 describe(assertPlanIsConsistent, () => {
   it('accepts a plan whose references, blobs, and statuses all agree', () => {
     expect(() => {
@@ -12,76 +16,13 @@ describe(assertPlanIsConsistent, () => {
     }).not.toThrow();
   });
 
-  it('if an edge names an artifact no table carries, locates the dangling reference', () => {
-    const plan = buildPlan();
-    requireEntry(plan.artifacts, 1).dependsOn = [{ to: 'skill:absent', via: 'declared' }];
-
-    expect(findViolations(plan)).toStrictEqual([
-      {
-        path: 'artifacts[1].dependsOn[0].to',
-        message: 'references "skill:absent", which is not an entry in artifacts',
-      },
-    ]);
-  });
-
-  it('if a file contributor names an unknown partial, locates the dangling reference', () => {
-    const plan = buildPlan();
-    requireEntry(plan.files, 0).contributors.partials = ['team:_data/absent.md'];
-
-    expect(findViolations(plan)).toStrictEqual([
-      {
-        path: 'files[0].contributors.partials[0]',
-        message: 'references "team:_data/absent.md", which is not an entry in partials',
-      },
-    ]);
-  });
-
-  it('if a seed names a tier no table carries, locates the dangling reference', () => {
-    const plan = buildPlan();
-    plan.tiers = [];
-
-    expect(findViolations(plan)).toStrictEqual([
-      {
-        path: 'artifacts[0].seededBy[0].tierId',
-        message: 'references "project", which is not an entry in tiers',
-      },
-    ]);
-  });
-
   it('if the tier table carries one id twice, names the repeated id', () => {
     const plan = buildPlan();
     plan.tiers = [...plan.tiers, { id: 'project', label: 'Project, again' }];
 
-    expect(findViolations(plan)).toStrictEqual([{ path: 'tiers', message: 'carries "project" more than once' }]);
-  });
-
-  it('if a table carries one id twice, names the repeated id', () => {
-    const plan = buildPlan();
-    plan.sources = [...plan.sources, { id: 'team', name: 'team-again', origin: { kind: 'directory', location: '/x' } }];
-
-    expect(findViolations(plan)).toStrictEqual([{ path: 'sources', message: 'carries "team" more than once' }]);
-  });
-
-  it('if two file entries claim one destination, names the repeated path', () => {
-    const plan = buildPlan();
-    plan.files = [...plan.files, { ...requireEntry(plan.files, 0) }];
-
-    expect(findViolations(plan)).toStrictEqual([
-      { path: 'files[1]', message: 'repeats the destination "skills/review/SKILL.md" within target "claude"' },
+    expect(captureFailure(plan).violations).toStrictEqual([
+      { path: 'tiers', message: 'carries "project" more than once' },
     ]);
-  });
-
-  it('tolerates one path claimed once per target', () => {
-    const plan = buildPlan();
-    plan.targets = [
-      ...plan.targets,
-      { id: 'rovodev', label: 'Rovo Dev', root: '~/.rovodev', tokenMappings: [], variables: [] },
-    ];
-    plan.files = [...plan.files, { ...requireEntry(plan.files, 0), targetId: 'rovodev' }];
-
-    expect(() => {
-      assertPlanIsConsistent(plan);
-    }).not.toThrow();
   });
 
   it('if a shadowed candidate outranks its winner, rejects the resolution order', () => {
@@ -92,33 +33,10 @@ describe(assertPlanIsConsistent, () => {
       shadowed: [{ sourceId: 'team', path: 'skills/lint/SKILL.md', hash: 'hash:lint-team' }],
     };
 
-    expect(findViolations(plan)).toStrictEqual([
+    expect(captureFailure(plan).violations).toStrictEqual([
       {
         path: 'artifacts[2].resolution.shadowed[0].sourceId',
         message: 'names "team", which does not follow "library" in source precedence order',
-      },
-    ]);
-  });
-
-  it('if shadowed candidates run against precedence, rejects the one out of order', () => {
-    const plan = buildPlan();
-    plan.sources = [
-      ...plan.sources,
-      { id: 'packaged', name: 'packaged', origin: { kind: 'package', location: '@acme/guidance' } },
-    ];
-    const review = requireEntry(plan.artifacts, 1);
-    review.resolution = {
-      winner: { sourceId: 'team', path: 'skills/review/SKILL.md', hash: 'hash:review' },
-      shadowed: [
-        { sourceId: 'packaged', path: 'skills/review/SKILL.md', hash: 'hash:review-packaged' },
-        { sourceId: 'library', path: 'skills/review/SKILL.md', hash: 'hash:review-library' },
-      ],
-    };
-
-    expect(findViolations(plan)).toStrictEqual([
-      {
-        path: 'artifacts[1].resolution.shadowed[1].sourceId',
-        message: 'names "library", which does not follow "packaged" in source precedence order',
       },
     ]);
   });
@@ -130,115 +48,11 @@ describe(assertPlanIsConsistent, () => {
       shadowed: [{ sourceId: 'absent', path: 'skills/review/SKILL.md', hash: 'hash:review-absent' }],
     };
 
-    expect(findViolations(plan)).toStrictEqual([
+    expect(captureFailure(plan).violations).toStrictEqual([
       {
         path: 'artifacts[1].resolution.shadowed[0].sourceId',
         message: 'references "absent", which is not an entry in sources',
       },
-    ]);
-  });
-
-  it('tolerates a removed artifact that no longer resolves from any source', () => {
-    const plan = buildPlan();
-    const retired = requireEntry(plan.artifacts, 2);
-    plan.artifacts = [...plan.artifacts, { id: retired.id, kindId: 'skill', slug: 'lint', status: 'removed' }];
-    plan.artifacts = plan.artifacts.filter((artifact) => artifact !== retired);
-
-    expect(() => {
-      assertPlanIsConsistent(plan);
-    }).not.toThrow();
-  });
-
-  it('blames each out-of-order candidate on the source it was compared against', () => {
-    const plan = buildPlan();
-    plan.sources = [
-      ...plan.sources,
-      { id: 'packaged', name: 'packaged', origin: { kind: 'package', location: '@acme/guidance' } },
-      { id: 'vendored', name: 'vendored', origin: { kind: 'directory', location: '/srv/vendored' } },
-    ];
-    // `vendored` is lowest, so both candidates after it are out of order against it, not against each other.
-    requireEntry(plan.artifacts, 1).resolution = {
-      winner: { sourceId: 'team', path: 'skills/review/SKILL.md', hash: 'hash:review' },
-      shadowed: [
-        { sourceId: 'vendored', path: 'skills/review/SKILL.md', hash: 'hash:review-vendored' },
-        { sourceId: 'packaged', path: 'skills/review/SKILL.md', hash: 'hash:review-packaged' },
-        { sourceId: 'library', path: 'skills/review/SKILL.md', hash: 'hash:review-library' },
-      ],
-    };
-
-    expect(findViolations(plan)).toStrictEqual([
-      {
-        path: 'artifacts[1].resolution.shadowed[1].sourceId',
-        message: 'names "packaged", which does not follow "vendored" in source precedence order',
-      },
-      {
-        path: 'artifacts[1].resolution.shadowed[2].sourceId',
-        message: 'names "library", which does not follow "vendored" in source precedence order',
-      },
-    ]);
-  });
-
-  it('if a non-token edge names a partial, rejects the edge it could not have been read from', () => {
-    const plan = buildPlan();
-    requireEntry(plan.artifacts, 0).dependsOn = [
-      { to: 'skill:review', via: 'member', partialId: 'team:_data/shared.md' },
-    ];
-
-    expect(findViolations(plan)).toStrictEqual([
-      {
-        path: 'artifacts[0].dependsOn[0].partialId',
-        message: 'is set on a "member" edge, and only a token edge is read from a partial',
-      },
-    ]);
-  });
-
-  it('if a plan claiming complete content omits a referenced body, names the missing hash', () => {
-    const plan = buildPlan();
-    plan.blobs = {};
-
-    expect(findViolations(plan)).toStrictEqual([
-      { path: 'files[0].current.hash', message: 'names "hash:review-current", which blobs does not carry' },
-      { path: 'files[0].planned.hash', message: 'names "hash:review-planned", which blobs does not carry' },
-    ]);
-  });
-
-  it('if a plan declares partial content, tolerates a body it does not carry', () => {
-    const plan = buildPlan();
-    plan.contentAvailability = 'partial';
-    plan.blobs = {};
-
-    expect(() => {
-      assertPlanIsConsistent(plan);
-    }).not.toThrow();
-  });
-
-  it('if both sides carry the same hash, rejects a status of changed', () => {
-    const plan = buildPlan();
-    requireEntry(plan.files, 0).planned = { hash: 'hash:review-current' };
-
-    expect(findViolations(plan)).toStrictEqual([
-      { path: 'files[0].status', message: 'is "changed", but its sides describe "unchanged"' },
-    ]);
-  });
-
-  it('if a file carries neither side, rejects it', () => {
-    const plan = buildPlan();
-    delete requireEntry(plan.files, 0).current;
-    delete requireEntry(plan.files, 0).planned;
-
-    expect(findViolations(plan)).toStrictEqual([
-      { path: 'files[0]', message: 'carries neither a current nor a planned side' },
-    ]);
-  });
-
-  it('if an unchanged file records a block, rejects the block that could not apply anyway', () => {
-    const plan = buildPlan();
-    requireEntry(plan.files, 0).status = 'unchanged';
-    requireEntry(plan.files, 0).planned = { hash: 'hash:review-current' };
-    requireEntry(plan.files, 0).blocked = { reason: 'the region host is malformed' };
-
-    expect(findViolations(plan)).toStrictEqual([
-      { path: 'files[0].blocked', message: 'is set on a file that would not be written anyway' },
     ]);
   });
 
@@ -248,19 +62,29 @@ describe(assertPlanIsConsistent, () => {
     requireEntry(plan.files, 0).planned = { hash: 'hash:review-current' };
     plan.blobs = {};
 
-    expect(findViolations(plan)).toHaveLength(4);
+    expect(captureFailure(plan).violations).toHaveLength(4);
+  });
+
+  it('raises a failure a consumer can catch alongside a catalog failure, under its own name', () => {
+    const plan = buildPlan();
+    plan.tiers = [];
+    const failure = captureFailure(plan);
+
+    expect(failure).toBeInstanceOf(ConsistencyError);
+    expect(failure.name).toBe('PlanConsistencyError');
+    expect(failure.message).toMatch(/^Plan is inconsistent:\n/);
   });
 });
 
 // region | Helpers
 
-/** The violations `plan` raises, failing the test when it is consistent. */
-function findViolations(plan: Plan): ReadonlyArray<PlanViolation> {
+/** The failure `plan` raises, failing the test when it passes every check. */
+function captureFailure(plan: Plan): PlanConsistencyError {
   try {
     assertPlanIsConsistent(plan);
   } catch (error: unknown) {
     if (error instanceof PlanConsistencyError) {
-      return error.violations;
+      return error;
     }
     throw error;
   }
