+++ packages/compositor/src/plan/checks/__tests__/findMisplacedPartialReferences.unit.test.ts
@@ -1,0 +1,25 @@
+import { describe, expect, it } from 'vitest';
+
+import { buildPlan } from '../../../test-utils/buildPlan.ts';
+import { requireEntry } from '../../../test-utils/requireEntry.ts';
+import { findMisplacedPartialReferences } from '../findMisplacedPartialReferences.ts';
+
+describe(findMisplacedPartialReferences, () => {
+  it('accepts a token edge naming the partial it was read from', () => {
+    expect(findMisplacedPartialReferences(buildPlan())).toStrictEqual([]);
+  });
+
+  it('if a non-token edge names a partial, rejects the edge it could not have been read from', () => {
+    const plan = buildPlan();
+    requireEntry(plan.artifacts, 0).dependsOn = [
+      { to: 'skill:review', via: 'member', partialId: 'team:_data/shared.md' },
+    ];
+
+    expect(findMisplacedPartialReferences(plan)).toStrictEqual([
+      {
+        path: 'artifacts[0].dependsOn[0].partialId',
+        message: 'is set on a "member" edge, and only a token edge is read from a partial',
+      },
+    ]);
+  });
+});
