+++ packages/compositor/src/plan/checks/findMissingBlobs.ts
@@ -1,0 +1,24 @@
+import type { Violation } from '../../consistency/Violation.ts';
+import type { Plan } from '../../schemas/plan-schema.ts';
+
+/** Violations for each file body a plan claiming complete content does not carry. */
+export function findMissingBlobs(plan: Plan): Array<Violation> {
+  if (plan.contentAvailability !== 'complete') {
+    return [];
+  }
+
+  const carried = new Set(Object.keys(plan.blobs));
+  const violations: Array<Violation> = [];
+  for (const [index, file] of plan.files.entries()) {
+    for (const side of ['current', 'planned'] as const) {
+      const hash = side === 'current' ? file.current?.hash : file.planned?.hash;
+      if (hash !== undefined && !carried.has(hash)) {
+        violations.push({
+          path: `files[${index}].${side}.hash`,
+          message: `names "${hash}", which blobs does not carry`,
+        });
+      }
+    }
+  }
+  return violations;
+}
